requirements.txt
PyYAML==5.4.1
docker==6.0.0